pyproject.toml
[tool.poetry]
name = "jafx"
version = "0.1.0"
description = "Composable deep learning with effectful JAX primitives"
authors = []

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
cattrs = "^1.8.0"
pendulum = "^2.1.2"
ansicolors = "^1.1.8"
dm-haiku = {git = "https://github.com/deepmind/dm-haiku.git", rev = "main"}
tensorboard = "^2.7.0"
tensorflow = "^2.6.1"
matplotlib = "^3.4.3"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
pytest = "^6.2.5"
black = "21.7b0"
isort = "^5.9.3"

[tool.isort]
profile = "black"
multi_line_output = 3
skip = "__init__.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
